LPS-53238 Move to separate upgrade process because it was added to UpgradeProcess_1_0_0 after its original introduction, so its possible that some users ran the UpgradeProcess_1_0_0 without UpgradePortletPreferences, in which case UpgradePortletPreferences will never run
Conflicts:
	portlets/calendar-portlet/docroot/WEB-INF/src/com/liferay/calendar/hook/upgrade/UpgradeProcess_1_0_1.java
