+++ portlets/calendar-portlet/docroot/WEB-INF/src/com/liferay/calendar/hook/upgrade/v1_0_2/UpgradePortletPreferences.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.calendar.hook.upgrade.v1_0_0;
+package com.liferay.calendar.hook.upgrade.v1_0_2;
 
 import com.liferay.calendar.model.CalendarBooking;
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
